Split MapService
