+++ src/components/map/PreviewLayersService.js
@@ -1,12 +1,14 @@
 goog.provide('ga_previewlayers_service');
 
-goog.require('ga_map_service');
+goog.require('ga_layers_service');
+goog.require('ga_maputils_service');
 goog.require('ga_wms_service');
 goog.require('ga_wmts_service');
 (function() {
 
   var module = angular.module('ga_previewlayers_service', [
-    'ga_map_service',
+    'ga_maputils_service',
+    'ga_layers_service',
     'ga_wms_service',
     'ga_wmts_service'
   ]);
